Vizality Support Crippled - READ MORE
I have not used Vizality for the better part of a year, and therefore have not kept up with the proper variables to give Vizality's settings the coverage it would need. For this very reason, I believe that it would be wise to remove what I consider to be "bloat". 

If you would like to add this back, you may add what's in the PasteBin link to the additions part of SlidetoUnlock.theme.css. https://pastebin.com/wTy8XHHS 

When all of the changes I've been making here are considered antiquate enough, I will be pushing this all to stable. When that happens, the Vizality manifest file will be removed. You may choose to make your own manifest file for Vizality if you please.